input validating: date pattern; unit tests for calculating versionValidUntil
