+++ src/utils/strings.js
@@ -34,4 +34,6 @@ export const eu = (dateString) => {
         return null;
     }
     return new Date(dateString)?.toLocaleString('ru-RU')?.substr(0, 10);
-}
+};
+
+export const datePattern = /^((?:18|19|20|21|22)\d{2})-((?:0[1-9])|(?:1[0-2]))-((?:0[0-9])|(?:[1-2][0-9])|(?:3[0-1]))T\d{2}:\d{2}:\d{2}.\d{3}Z$/i //2020-09-21T00:00:00.000Z
